Split tests for lattice and stream, and fix test_common reprs
